third_party/blink/web_tests/platform/win/external/wpt/editing/run/bold_1001-2000-expected.txt

This is a testharness.js-based test.
Found 152 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandState("bold") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar]</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: bold\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: bold\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p><span style=\\"font-weight:bold\\">baz</span></p>" but got "<p style=\\"\\">foo</p><p style=\\"\\">bar</p><p style=\\"font-weight:bold\\">baz</p>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p><b>baz</b></p>" but got "<p style=\\"\\">foo</p><p style=\\"\\">bar</p><p style=\\"font-weight:bold\\">baz</p>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b>{<p>foo</p><p>bar</p>}<p>baz</p></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-weight:bold\\">foo</span><i>bar</i></p><p><span style=\\"font-weight:bold\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-weight:bold\\">foo</span><i style=\\"\\">bar</i></p><p style=\\"font-weight:bold\\">baz</p>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foo</b><i>bar</i></p><p><b>baz</b></p>" but got "<p style=\\"\\"><b>foo</b><i style=\\"\\">bar</i></p><p style=\\"font-weight:bold\\">baz</p>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b><p>foo[<i>bar</i>}</p><p>baz</p></b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span id=\\"purple\\"><span style=\\"font-weight:bold\\">bar </span>baz<span style=\\"font-weight:bold\\"> qoz</span></span>" but got "<b id=\\"purple\\">bar </b>baz<b> qoz</b>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span id=\\"purple\\"><b>bar </b>baz<b> qoz</b></span>" but got "<b id=\\"purple\\">bar </b>baz<b> qoz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<b id=purple>bar [baz] qoz</b>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 100\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 100\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 100\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 100\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 200\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 200\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 200\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 200\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 300\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 300\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 300\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 300\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 500\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 500\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 500\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 500\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 600\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 600\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 600\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 600\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 800\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 800\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 800\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 800\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 900\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 900\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 900\\">[bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 900\\">[bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 400\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo<span style=\\"font-weight: 700\\">[bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar]</span>baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar]</span>baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 400\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar</span>]baz" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "foo[<span style=\\"font-weight: 700\\">bar</span>]baz" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">foo<span style=\\"font-weight:bold\\">bar</span>baz</span>" but got "<span style=\\"font-weight:100\\">foo</span><span style=\\"font-weight:bold\\">bar</span><span style=\\"font-weight:100\\">baz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">foo<b>bar</b>baz</span>" but got "<span style=\\"font-weight:100\\">foo</span><b>bar</b><span style=\\"font-weight:100\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">foo<span style=\\"font-weight:bold\\">bar</span>baz</span>" but got "<span style=\\"font-weight:400\\">foo</span><span style=\\"font-weight:bold\\">bar</span><span style=\\"font-weight:400\\">baz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">foo<b>bar</b>baz</span>" but got "<span style=\\"font-weight:400\\">foo</span><b>bar</b><span style=\\"font-weight:400\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[bar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo</b>bar<b>baz</b>" but got "<span style=\\"font-weight:700\\">foo</span>bar<span style=\\"font-weight:700\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 900\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 900\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 900\\">foo[bar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 900\\">foo[bar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\"><span style=\\"font-weight:bold\\">foobar</span>baz</span>" but got "<span style=\\"font-weight:bold\\">foobar</span><span style=\\"font-weight:100\\">baz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\"><b>foobar</b>baz</span>" but got "<b>foobar</b><span style=\\"font-weight:100\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 100\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\"><span style=\\"font-weight:bold\\">foobar</span>baz</span>" but got "<span style=\\"font-weight:bold\\">foobar</span><span style=\\"font-weight:400\\">baz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\"><b>foobar</b>baz</span>" but got "<b>foobar</b><span style=\\"font-weight:400\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 400\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 700\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "{<span style=\\"font-weight: 700\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 700\\">foobar]baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar<b>baz</b>" but got "foobar<span style=\\"font-weight:700\\">baz</span>"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 700\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "{<span style=\\"font-weight: 700\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""]] "{<span style=\\"font-weight: 900\\">foobar]baz</span>" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""]] "{<span style=\\"font-weight: 900\\">foobar]baz</span>" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">foo<span style=\\"font-weight:bold\\">barbaz</span></span>" but got "<span style=\\"font-weight:100\\">foo</span><span style=\\"font-weight:bold\\">barbaz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:100\\">foo<b>barbaz</b></span>" but got "<span style=\\"font-weight:100\\">foo</span><b>barbaz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 100\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">foo<span style=\\"font-weight:bold\\">barbaz</span></span>" but got "<span style=\\"font-weight:400\\">foo</span><span style=\\"font-weight:bold\\">barbaz</span>"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-weight:400\\">foo<b>barbaz</b></span>" but got "<span style=\\"font-weight:400\\">foo</span><b>barbaz</b>"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 400\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[barbaz</span>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo</b>barbaz" but got "<span style=\\"font-weight:700\\">foo</span>barbaz"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[barbaz</span>}" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["bold",""]] "<span style=\\"font-weight: 700\\">foo[barbaz</span>}" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
